requirements.txt
cycler==0.10.0
drawnow==0.71.3
matplotlib==2.1.0
numpy==1.13.3
pyparsing==2.2.0
pyserial==3.4
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
